Reduce size of trace record type

+++ trace/simple.c
@@ -162,7 +162,7 @@ static gpointer writeout_thread(gpointer opaque)
     unsigned int idx = 0;
     //int dropped_count;
     size_t unused __attribute__ ((unused));
-    uint64_t type = TRACE_RECORD_TYPE_EVENT;
+    uint8_t type = TRACE_RECORD_TYPE_EVENT;
 
     for (;;) {
         wait_for_trace_records_available();
@@ -288,7 +288,7 @@ void trace_record_finish(TraceBufferRecord *rec)
 
 static int st_write_event_mapping(void)
 {
-    uint64_t type = TRACE_RECORD_TYPE_MAPPING;
+    uint8_t type = TRACE_RECORD_TYPE_MAPPING;
     TraceEventIter iter;
     TraceEvent *ev;
